Correção: Melhorias na interface e alterações básicas na funcionalidade do método de formatação de títulos.

+++ src/main.py
@@ -22,22 +22,14 @@ class App:
         my_font = customtkinter.CTkFont(family="Segoe UI", size=14, weight="bold")
         try:
             def progress_hook(d):
-                # Atualiza o progresso do download
                 if d['status'] == 'downloading' and progress_bar:
                     downloaded = d.get('downloaded_bytes', 0)
                     total = d.get('total_bytes', None)
-
-                    if total and total > 0:  # Verifica se 'total_bytes' é válido
-                        progress = downloaded / total
-                    else:
-                        progress = 0  # Define progresso como 0 se não for possível calcular
-
+                    progress = downloaded / total if total else 0
                     progress_bar.set(progress)
-                    progress_label.configure(text=f"Baixando... {int(progress * 100)}%", font=my_font)
-
-                # Indica que o download foi finalizado
+                    progress_label.configure(text=f"Baixando... {int(progress * 100)}%")
                 elif d['status'] == 'finished':
-                    progress_label.configure(text="Download concluído, iniciando conversão...", font=my_font)
+                    progress_label.configure(text="Download concluído, iniciando conversão...")
 
             ydl_opts = {
                 'format': 'bestvideo[ext=mp4]+bestaudio[ext=m4a]/best[ext=mp4]',
@@ -50,50 +42,64 @@ class App:
                 info_dict = ydl.extract_info(self.link, download=True)
 
                 if isinstance(info_dict, dict) and 'entries' in info_dict:
-                    # É uma playlist
+                    total_videos = len(info_dict['entries'])
+                    downloaded_count = 0
                     for entry in info_dict['entries']:
-                        downloaded_file = ydl.prepare_filename(entry)
+                        try:
+                            if entry is None or 'title' not in entry:
+                                raise ValueError("Vídeo indisponível na playlist.")
+
+                            downloaded_file = ydl.prepare_filename(entry)
+                            if downloaded_file.endswith(".webm"):
+                                self.convert_to_mp4(downloaded_file, progress_bar, progress_label)
+                                os.remove(downloaded_file)
+
+                            final_file = downloaded_file.replace(".webm", ".mp4") if downloaded_file.endswith(".webm") else downloaded_file
+
+                            if self.formatar_nomes:
+                                novo_nome = self.formatar_nome(os.path.splitext(os.path.basename(final_file))[0], ".mp4")
+                                novo_caminho = os.path.join(self.destination, novo_nome)
+                                os.rename(final_file, novo_caminho)
+                                final_file = novo_caminho
+
+                            downloaded_count += 1
+                            progress_label.configure(text=f"Baixando... [{downloaded_count}/{total_videos}]")
+                        except ValueError as ve:
+                            print(f"Erro: {ve} - Ignorando entrada.")
+                        except Exception as e:
+                            print(f"Erro ao processar o vídeo da playlist: {e}")
+                else:
+                    try:
+                        if info_dict is None or 'title' not in info_dict or info_dict.get('is_private') or info_dict.get('was_live') or info_dict.get('uploader') is None:
+                            raise ValueError("Vídeo indisponível ou não encontrado.")
+
+                        downloaded_file = ydl.prepare_filename(info_dict)
                         if downloaded_file.endswith(".webm"):
                             self.convert_to_mp4(downloaded_file, progress_bar, progress_label)
                             os.remove(downloaded_file)
 
-                        # Formata o nome do arquivo
-                        final_file = downloaded_file.replace(".webm", ".mp4") if downloaded_file.endswith(
-                            ".webm") else downloaded_file
+                        final_file = downloaded_file.replace(".webm", ".mp4") if downloaded_file.endswith(".webm") else downloaded_file
 
                         if self.formatar_nomes:
                             novo_nome = self.formatar_nome(os.path.splitext(os.path.basename(final_file))[0], ".mp4")
                             novo_caminho = os.path.join(self.destination, novo_nome)
                             os.rename(final_file, novo_caminho)
                             final_file = novo_caminho
-                else:
-                    # Download único
-                    downloaded_file = ydl.prepare_filename(info_dict)
-                    if downloaded_file.endswith(".webm"):
-                        self.convert_to_mp4(downloaded_file, progress_bar, progress_label)
-                        os.remove(downloaded_file)
+                    except ValueError as ve:
+                        messagebox.showerror("Erro", f"{ve}")
+                        return
+                    except Exception as e:
+                        print(f"Erro ao processar o vídeo: {e}")
 
-                    # Formata o nome do arquivo
-                    final_file = downloaded_file.replace(".webm", ".mp4") if downloaded_file.endswith(
-                        ".webm") else downloaded_file
-
-                    if self.formatar_nomes:
-                        novo_nome = self.formatar_nome(os.path.splitext(os.path.basename(final_file))[0], ".mp4")
-                        novo_caminho = os.path.join(self.destination, novo_nome)
-                        os.rename(final_file, novo_caminho)
-                        final_file = novo_caminho
-
-            # Verifica se o arquivo final existe
             if not os.path.exists(final_file):
                 raise FileNotFoundError(f"O arquivo final {final_file} não foi gerado.")
 
             if progress_window:
-                # Exibe informações finais do download
                 for widget in progress_window.winfo_children():
                     widget.destroy()
                 absolute_path = os.path.abspath(final_file)
                 local_path = os.path.dirname(absolute_path)
-                progress_label = ctk.CTkLabel(
+                progress_label = customtkinter.CTkLabel(
                     progress_window,
                     text=f"Download concluído!\n\nArquivo: {os.path.basename(final_file)}\nLocal: {local_path}",
                     font=my_font,
@@ -101,7 +107,7 @@ class App:
                 )
                 progress_label.pack(pady=20)
 
-                close_button = ctk.CTkButton(
+                close_button = customtkinter.CTkButton(
                     progress_window,
                     text="Fechar",
                     command=progress_window.destroy,
@@ -120,6 +126,119 @@ class App:
             if progress_window:
                 progress_window.destroy()
             messagebox.showerror("Erro", f"Ocorreu um erro ao baixar ou converter o vídeo: {e}")
+
+    def download_mp3(self, progress_bar=None, progress_label=None, progress_window=None):
+        my_font = customtkinter.CTkFont(family="Segoe UI", size=14, weight="bold")
+        try:
+            def progress_hook(d):
+                # Atualiza o progresso do download
+                if d['status'] == 'downloading' and progress_bar:
+                    downloaded = d.get('downloaded_bytes', 0)
+                    total = d.get('total_bytes', None)
+
+                    if total and total > 0:  # Verifica se 'total_bytes' é válido
+                        progress = downloaded / total
+                    else:
+                        progress = 0  # Define progresso como 0 se não for possível calcular
+
+                    progress_bar.set(progress)
+                    progress_label.configure(text=f"Baixando... {int(progress * 100)}%", font=my_font)
+
+                # Indica que o download foi finalizado
+                elif d['status'] == 'finished':
+                    progress_label.configure(text="Download concluído!", font=my_font)
+
+            ydl_opts = {
+                'format': 'bestaudio/best',  # Obtém o melhor áudio disponível
+                'outtmpl': os.path.join(self.destination, '%(title)s.%(ext)s'),
+                'postprocessors': [{
+                    'key': 'FFmpegExtractAudio',
+                    'preferredcodec': 'mp3',
+                    'preferredquality': '192',
+                }],
+                'ffmpeg_location': self.ffmpeg_path,
+                'progress_hooks': [progress_hook]
+            }
+
+            with YoutubeDL(ydl_opts) as ydl:
+                info_dict = ydl.extract_info(self.link, download=True)
+
+                # Processa os nomes dos arquivos baixados
+                if isinstance(info_dict, dict) and 'entries' in info_dict:
+                    # É uma playlist
+                    total_videos = len(info_dict['entries'])
+                    downloaded_count = 0
+                    for entry in info_dict['entries']:
+                        try:
+                            if entry is None or 'title' not in entry:
+                                raise ValueError("Vídeo indisponível na playlist.")
+
+                            downloaded_file = ydl.prepare_filename(entry).replace(".webm", ".mp3")
+                            final_file = downloaded_file
+                            if self.formatar_nomes:
+                                novo_nome = self.formatar_nome(os.path.splitext(os.path.basename(downloaded_file))[0],
+                                                               ".mp3")
+                                novo_caminho = os.path.join(self.destination, novo_nome)
+                                os.rename(downloaded_file, novo_caminho)
+                                final_file = novo_caminho
+
+                            downloaded_count += 1
+                            progress_label.configure(text=f"Baixando... [{downloaded_count}/{total_videos}]", font=my_font)
+                        except ValueError as ve:
+                            print(f"Erro: {ve} - Ignorando entrada.")
+                        except Exception as e:
+                            print(f"Erro ao processar o vídeo da playlist: {e}")
+                else:
+                    try:
+                        if info_dict is None or 'title' not in info_dict or info_dict.get('is_private') or info_dict.get('was_live') or info_dict.get('uploader') is None:
+                            raise ValueError("Vídeo indisponível ou não encontrado.")
+
+                        downloaded_file = ydl.prepare_filename(info_dict).replace(".webm", ".mp3")
+                        final_file = downloaded_file
+                        if self.formatar_nomes:
+                            novo_nome = self.formatar_nome(os.path.splitext(os.path.basename(downloaded_file))[0], ".mp3")
+                            novo_caminho = os.path.join(self.destination, novo_nome)
+                            os.rename(downloaded_file, novo_caminho)
+                            final_file = novo_caminho
+                    except ValueError as ve:
+                        messagebox.showerror("Erro", f"{ve}")
+                        return
+                    except Exception as e:
+                        print(f"Erro ao processar o vídeo: {e}")
+
+            if progress_window:
+                for widget in progress_window.winfo_children():
+                    widget.destroy()
+                absolute_path = os.path.abspath(final_file)
+                local_path = os.path.dirname(absolute_path)
+                progress_label = customtkinter.CTkLabel(
+                    progress_window,
+                    text=f"Download concluído!\n\nArquivo: {os.path.basename(final_file)}\nLocal: {local_path}",
+                    font=my_font,
+                    justify="center"
+                )
+                progress_label.pack(pady=20)
+
+                close_button = customtkinter.CTkButton(
+                    progress_window,
+                    text="Fechar",
+                    command=progress_window.destroy,
+                    width=110,
+                    height=40,
+                    corner_radius=50,
+                    fg_color="#c74066",
+                    font=my_font,
+                    hover_color="#b03558",
+                    border_color="#000000",
+                    border_width=1
+                )
+                close_button.pack(pady=10)
+
+        except Exception as e:
+            if progress_window:
+                progress_window.destroy()
+            messagebox.showerror("Erro", f"Ocorreu um erro ao baixar o áudio: {e}")
+
 
     def convert_to_mp4(self, input_file, progress_bar=None, progress_label=None):
         my_font = customtkinter.CTkFont(family="Segoe UI", size=14, weight="bold")
@@ -225,28 +344,31 @@ class App:
                     # É uma playlist
                     for entry in info_dict['entries']:
                         downloaded_file = ydl.prepare_filename(entry).replace(".webm", ".mp3")
+                        final_file = downloaded_file
                         if self.formatar_nomes:
                             novo_nome = self.formatar_nome(os.path.splitext(os.path.basename(downloaded_file))[0],
                                                            ".mp3")
                             novo_caminho = os.path.join(self.destination, novo_nome)
                             os.rename(downloaded_file, novo_caminho)
+                            final_file = novo_caminho
                 else:
                     # Download único
                     downloaded_file = ydl.prepare_filename(info_dict).replace(".webm", ".mp3")
+                    final_file = downloaded_file
                     if self.formatar_nomes:
                         novo_nome = self.formatar_nome(os.path.splitext(os.path.basename(downloaded_file))[0], ".mp3")
                         novo_caminho = os.path.join(self.destination, novo_nome)
                         os.rename(downloaded_file, novo_caminho)
-                        downloaded_file = novo_caminho
+                        final_file = novo_caminho
 
             if progress_window:
                 for widget in progress_window.winfo_children():
                     widget.destroy()
-                absolute_path = os.path.abspath(downloaded_file)
+                absolute_path = os.path.abspath(final_file)
                 local_path = os.path.dirname(absolute_path)
                 progress_label = ctk.CTkLabel(
                     progress_window,
-                    text=f"Download concluído!\n\nArquivo: {os.path.basename(downloaded_file)}\nLocal: {local_path}",
+                    text=f"Download concluído!\n\nArquivo: {os.path.basename(final_file)}\nLocal: {local_path}",
                     font=my_font,
                     justify="center"
                 )
@@ -387,7 +509,6 @@ class BaixarVideo:
         )
         self.checkbox.grid(row=0, column=0, padx=(0, 5), sticky="w")
 
-        # Asterisco com Tooltip (ao lado da label da checkbox)
         self.tooltip_label = ctk.CTkLabel(self.checkbox_frame, text="*", text_color="#c74066", cursor="hand2")
         self.tooltip_label.grid(row=0, column=1, sticky="w", padx=1)
         self.tooltip_label.bind("<Enter>", lambda e: self.show_tooltip(e,
@@ -404,7 +525,14 @@ class BaixarVideo:
         self.dest_path = ctk.CTkLabel(self.left_frame,
                                       text="Diretório Indefinido",
                                       text_color="white")
-        self.dest_path.grid(row=3, column=2, columnspan=2, padx=10, pady=5, sticky="w")
+        self.dest_path.grid(row=3, column=2, columnspan=2, padx=20, pady=5, sticky="w")
+        #Tooltip
+        self.tooltip_label = ctk.CTkLabel(self.left_frame, text="*", text_color="#c74066", cursor="hand2")
+        self.tooltip_label.grid(row=3, column=2, columnspan =2, sticky="w", padx=10, pady=5)
+        self.tooltip_label.bind("<Enter>", lambda e: self.show_tooltip(e,
+                                                                           "Ao clicar no botão:\n- Altera o diretório em que o arquivo será baixado\n- Caso não selecione um, irá para o diretório padrão do programa.",
+                                                                       x_offset=-180, y_offset=-70))
+        self.tooltip_label.bind("<Leave>", self.hide_tooltip)
 
         # Frame direito para informações do vídeo
         self.info_frame = ctk.CTkFrame(root, corner_radius=10)
@@ -502,7 +630,7 @@ class BaixarVideo:
                 self.thumbnail_label.configure(text=info['title'])
                 details = (f"Duração: {minutes}:{seconds:02}\n"
 
-                           f"Canal: {info['channel']}\n\n"
+                           f"Canal: {info['channel']}\n"
 
                            f"Descrição:\n{info['description'][:70]}...")
 
@@ -512,7 +640,7 @@ class BaixarVideo:
                         response = requests.get(info['thumbnail'], stream=True)
                         if response.status_code == 200:
                             img = Image.open(BytesIO(response.content))
-                            thumbnail = ctk.CTkImage(light_image=img, size=(320, 180))
+                            thumbnail = ctk.CTkImage(light_image=img,size=(320, 180))
 
                             self.thumbnail_image.configure(image=thumbnail)
                             self.thumbnail_image.image = thumbnail
